+++ .idea/EasyDrill.iml
@@ -2,7 +2,6 @@
 <module type="PYTHON_MODULE" version="4">
   <component name="NewModuleRootManager">
     <content url="file://$MODULE_DIR$">
-      <sourceFolder url="file://$MODULE_DIR$/Modelo/Tuberia" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/Modelo" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/Modelo/Hidraulica/Metodos" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/Modelo/Hidraulica" isTestSource="false" />
@@ -14,6 +13,9 @@
       <sourceFolder url="file://$MODULE_DIR$/Controlador/Hidraulica" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/Modelo/Direccional" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/Controlador" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/Recursos" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/Recursos/Constantes" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/Recursos/Pruebas" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/venv" />
     </content>
     <orderEntry type="jdk" jdkName="Python 3.7 (Tesis)" jdkType="Python SDK" />
