--- Modelo/Tuberia/Exterior.py
@@ -1,11 +1,0 @@
-from Tubo import *
-
-
-class Exterior (General):
-    Tipo = ""
-    BocaL = False
-    CAnular = 0
-    VolAnular = 0
-    tinterior = ""
-
-
